Improved RigidBody - Level interaction

+++ src/RigidBody/RigidBody.hpp
@@ -9,10 +9,10 @@ struct RigidBody {
 public:
     int id;
     vec2 position;
-    int mass;
+    float mass;
     float rotation;
 
-    RigidBody(int id, float x, float y, int mass = 1, float rotation = 0.f)
+    RigidBody(int id, float x, float y, float mass = 1.f, float rotation = 0.f)
         : id(id), position(x, y), mass(mass), rotation(rotation) {}
     virtual ~RigidBody() {}
 };
@@ -21,7 +21,7 @@ struct RigidBodyRect : public RigidBody {
     float width;
     float height;
 
-    RigidBodyRect(int id, float x, float y, int width, int height, int mass = 1,
+    RigidBodyRect(int id, float x, float y, int width, int height, float mass = 1.f,
                   float rotation = 0.f)
         : RigidBody(id, x, y, mass, rotation), width(width), height(height) {}
 };
@@ -29,7 +29,7 @@ struct RigidBodyRect : public RigidBody {
 struct RigidBodyCircle : public RigidBody {
     float radius;
 
-    RigidBodyCircle(int id, float x, float y, float radius = 1.f, int mass = 1,
+    RigidBodyCircle(int id, float x, float y, float radius = 1.f, float mass = 1.f,
                     float rotation = 0.f)
         : RigidBody(id, x, y, mass, rotation), radius(radius) {}
 };
@@ -37,7 +37,7 @@ struct RigidBodyCircle : public RigidBody {
 struct RigidBodyTriangle : public RigidBody {
     vec2 points[3];
 
-    RigidBodyTriangle(int id, float x, float y, vec2 p1, vec2 p2, vec2 p3, int mass = 1,
+    RigidBodyTriangle(int id, float x, float y, vec2 p1, vec2 p2, vec2 p3, float mass = 1.f,
                       float rotation = 0.f)
         : RigidBody(id, x, y, mass, rotation) {
         points[0] = p1;
@@ -45,7 +45,7 @@ struct RigidBodyTriangle : public RigidBody {
         points[2] = p3;
     }
 
-    RigidBodyTriangle(int id, float x, float y, vec2 p2, vec2 p3, int mass = 1,
+    RigidBodyTriangle(int id, float x, float y, vec2 p2, vec2 p3, float mass = 1.f,
                       float rotation = 0.f)
         : RigidBodyTriangle(id, x, y, vec2(0, 0), p2, p3, mass, rotation) {}
 };
